src/scraper/async_scraper.py
import asyncio
import httpx
from bs4 import BeautifulSoup
from typing import List, Dict
import logging

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class AsyncScraper:
    def __init__(self, headers: Dict = None):
        self.headers = headers or {
            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
        }
        self.client = httpx.AsyncClient(headers=self.headers, follow_redirects=True, timeout=30.0)

    async def fetch_page(self, url: str) -> str:
        """Fetch page content asynchronously."""
        try:
            response = await self.client.get(url)
            response.raise_for_status()
            return response.text
        except Exception as e:
            logger.error(f"Error fetching {url}: {e}")
            return ""

    async def scrape_multiple(self, urls: List[str]) -> List[str]:
        """Scrape multiple URLs concurrently."""
        tasks = [self.fetch_page(url) for url in urls]
        return await asyncio.gather(*tasks)

    def parse_articles(self, html: str, selector: str) -> List[Dict]:
        """Basic parser example."""
        soup = BeautifulSoup(html, 'html.parser')
        articles = []
        # Logic to be implemented based on specific news site structure
        return articles

    async def close(self):
        await self.client.aclose()

# Example usage pattern for a specific news site
class DinamalarScraper(AsyncScraper):
    async def get_latest_news(self):
        url = "https://www.dinamalar.com/tamil-news.asp"
        html = await self.fetch_page(url)
        # Parse logic here
        return []
